NSL-4523: Loader: Manually added excluded names are now flagged as Excluded in search results
In connection with this change I have renamed the column loader_name.excluded to loader_name.unplaced_flag.

This is to remove any confusion between a record_type of “excluded” and what was previously the column loader_name.excluded.

The only thing that now indicates a record is excluded is the record type of “excluded”.

Detailed changes

The form for creating a top-level loader_name record has changed:
   - you  have to choose a record type (accepted or excluded) and this is a mandatory field
   - there is  no check box for “unplaced”

I incidentally removed the distribution field from the form for editing excluded names.

Synonyms are no longer flagged as excluded synonyms in search results.