NSL-4713: Loader: Correct the query used on the loader_name detail tab to detect and warn about name-match-in-synonymy cases.
Added a new raw sql select based on the join query used to work out the name-match-in-syn: sub-query.

The query isn't crazy complicated, but getting it right (hopefully) took a lot of work - it relies on the tree_join_v which
abstracts the complexity of the tree tables into a single view, and that helps.  Decomposing it into an ActiveRecord query
seemed unhelpful.  If we need to work on it further in future the raw sql form will be easier to work with imo.