NSL-5180:  Loader: Remove the batch drop down from Loader 2 tab
The batch drop down was not being honoured by the typeahead dropdown because that typeahead is set up when the form is loaded
and therefore doesn't know what batch the user could have chosen when the form is running.

This made the batch dropdown misleading and unhelpful.

I chose to simply remove the batch dropdown and not go down the path of setting up JS code to reset the typeahead
 any time a user changes the batch selection in the form.  To take the more complex approach would be a departure from the
KISS approach to forms in the Editor.  There are so many forms and the goal is to keep them generally maintainable by keeping
them simple.

Also, it is very easy (two clicks) to reset the default batch from the menu.  I think that's a very easy alternative.

The users rep has allowed this approach as a trial.