NSL-5164:  Taxonomy: Honour tree.is_read_only
Prevent tree operations when boolean is set true.

Covers setting a working draft, editing a draft, publishing and draft and creating a draft.

Also many adjustments to the original tree code that did not follow Rails or Editor conventions.

Much more work to be done there but not via this Jira.