Build: Remove scripted changing of active_record timezome from :localhost to :utc - just leave production AR TZ as :utc
The config-prod.ed script is failing on teamcity with no clear message why.

The script simply changes the prod active_record TZ from :local to :utc.

I added this script when migrating from CSIRO to AWS - and the script was probably needed then during the
transition from one environment to the other.

We now don't need the script and will just leave the AR default TZ as :utc.